Cards in light theme

+++ app/src/main/java/com/supter/data/db/entity/PurchaseEntity.kt
@@ -18,7 +18,7 @@ data class PurchaseEntity(
     var price: Double,
     var order: Int,
     var stage: String,
-    var potential: Int,
+    var potential: Float,
     var description: String?,
     var questionsJSON: String?,
     var remind: Double,
